Use shutil.which to find executable
The existing method for searching PATH for the configured rustfmt binary
cause problems if you use the same config across platforms. On Windows,
the program needs to be "rustfmt.exe", not simply "rustfmt", whereas on
macOS or Linux, the opposite is the case. This means users can't share
config between those systems (by symlinking to Dropbox or similar).

Python 3 comes with shutil.which() which does everything we need here
including appending ".exe" on Windows if that's not already there.